refactor(daemon): тесты рантайма v5 — в отдельный файл
`runtime_v5.rs` перевалил за восемь тысяч строк только из-за модуля тестов:
production — 6311 строк, тесты — 2856. Модуль переехал в
`runtime_v5/tests.rs`; путь модуля прежний
(`infrastructure::daemon::runtime_v5::tests`), поэтому отбор nextest выбирает
те же 39 тестов. `include_str!` в тесте порядка полей смотрит на
`../runtime_v5.rs` — структура осталась в производственном файле.
